Move the ToUpperFirst method into the Library project.
