+++ BlazorSpark.Console/Commands/Pages/CreateComponentCommand.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using static System.Runtime.InteropServices.JavaScript.JSType;
+using BlazorSpark.Library.Extensions;
 
 namespace BlazorSpark.Console.Commands.Pages
 {
