+++ BlazorSpark.Console/Shared/StringExtensions.cs
@@ -14,22 +14,7 @@ namespace BlazorSpark.Console.Shared
             return Regex.Replace(value, @"(?<=[a-z])(?=[A-Z])", "-").ToLower();
         }
 
-        public static string ToUpperFirst(this string value)
-        {
-            if (string.IsNullOrEmpty(value))
-            {
-                return value;
-            }
 
-            if (value.Length == 1)
-            {
-                return value.ToUpper();
-            }
-
-            char firstChar = char.ToUpper(value[0]);
-            string restOfString = value[1..];
-            return firstChar + restOfString;
-        }
 
     }
 }
